+++ src/config/theme/osmosis.ts
@@ -19,20 +19,6 @@ const OSMOSIS_BASE_THEME: ThemeConfig = {
       },
     },
   },
-  fonts: {
-    heading: {
-      url: "https://fonts.googleapis.com/css2?family=Poppins:wght@500;600&display=swap",
-      name: "Poppins, serif",
-    },
-    body: {
-      url: "https://fonts.googleapis.com/css2?family=Inter:wght@300..700&display=swap",
-      name: "Inter, sans-serif",
-    },
-    mono: {
-      url: "/font/main.css",
-      name: "PP Neue Montreal Mono Regular",
-    },
-  },
   colors: {
     error: {
       main: "#FF666E",
@@ -194,6 +180,11 @@ const OSMOSIS_COMPONENT_CONFIG: ComponentConfig = {
     voted: "primary.main",
     votedAbstain: "primary.darker",
     didNotVote: "gray.700",
+  },
+  recentBlocks: {
+    signed: "primary.main",
+    proposed: "secondary.darker",
+    missed: "error.dark",
   },
 };