+++ src/config/theme/sei.ts
@@ -19,20 +19,6 @@ const SEI_BASE_THEME: ThemeConfig = {
       },
     },
   },
-  fonts: {
-    heading: {
-      url: "https://fonts.cdnfonts.com/css/satoshi?styles=135009,135005,135007,135002,135000",
-      name: "Satoshi, sans-serif",
-    },
-    body: {
-      url: "https://fonts.cdnfonts.com/css/satoshi?styles=135009,135005,135007,135002,135000",
-      name: "Satoshi, sans-serif",
-    },
-    mono: {
-      url: "/font/main.css",
-      name: "PP Neue Montreal Mono Regular",
-    },
-  },
   colors: {
     error: {
       main: "#FF666E",
@@ -111,7 +97,7 @@ const SEI_BASE_THEME: ThemeConfig = {
     },
   },
   footer: {
-    logo: "https://www.sei.io/_next/static/media/logo-light.1249fa55.svg",
+    logo: "https://assets.alleslabs.dev/integrations/sei/logo-sei.png",
     description:
       "A Smart Contract Explorer for Sei | Explore, deploy, execute, and query smart contracts on Sei from a user-friendly web UI",
     iconStyle: "rounded",
@@ -191,6 +177,11 @@ const SEI_COMPONENT_CONFIG: ComponentConfig = {
     voted: "accent.main",
     votedAbstain: "accent.darker",
     didNotVote: "gray.700",
+  },
+  recentBlocks: {
+    signed: "secondary.main",
+    proposed: "primary.main",
+    missed: "error.dark",
   },
 };