+++ src/config/theme/types.ts
@@ -13,20 +13,6 @@ export type ThemeConfig = {
       };
     };
   };
-  fonts: {
-    heading: {
-      url: string;
-      name: string;
-    };
-    body: {
-      url: string;
-      name: string;
-    };
-    mono: {
-      url: string;
-      name: string;
-    };
-  };
   colors: {
     primary: {
       main: string;
@@ -177,5 +163,10 @@ export type ComponentConfig = {
     voted: string;
     votedAbstain: string;
     didNotVote: string;
+  };
+  recentBlocks: {
+    signed: string;
+    proposed: string;
+    missed: string;
   };
 };