+++ src/lib/services/staking/api.ts
@@ -1,0 +1,10 @@
+import axios from "axios";
+
+import { zDelegationData } from "lib/services/types";
+import type { BechAddr } from "lib/types";
+import { parseWithError } from "lib/utils";
+
+export const getDelegationData = async (endpoint: string, address: BechAddr) =>
+  axios
+    .get(`${endpoint}/${encodeURI(address)}/delegations`)
+    .then(({ data }) => parseWithError(zDelegationData, data));